scd-hwmon: Add scd_smbus kernel trace support
Add kernel traces (/sys/kernel/debug/tracing) to the smbus component
of scd-hwmon.ko.

Creates a new header src/scd-smbus-trace.h. It needs to be an include,
due to the particular way kernel trace definitions are built (using
multiple inclusions).

Consequently, this change moves a lot of smbus-related type
definitions another header, scd-smbus.h, in order to share them. This
seems to align well with scd-mdio.h.

Hwmon smbus traces are basically register traces, i.e. the trace
buffer is effectively a sequence of 32-bit register values as written,
or observed over time, amended with context information like the SCD
pciId and accellerator number. It is expected to be used with
i2c/smbus core traces for the output to become more meaningful.

Users would typically enable traces in /sys/kernel/debug/tracing and
then either follow the trace output in realtime (for development), or
dump / snapshot the trace buffer on-demand (for debugging or
post-mortem analysis).

Change-Id: I700269e3335b0cd316a4d01c2d3d88de2fa44ba7