Enable PCI port only at the end of card power on
This change fixes outstanding issues with the SUP GMAC and PCI
strangeness leading to SUP lockup. The problematic state seems to be
caused by some sort of race condition at card power on on the PCI bus.
My best guess is that it happens when we setup the PLX and do a PCI
rescan. Sometimes it is possible that the tg3 driver is engaged when
the PCI tree is still in a "bogus" (non initialized) state. If this
happens, then at power off time, things will blow up.

In addition to the enable/disable change, it seems like a time.sleep(4)
is needed before linecard CPU can be deemed "powered on". This is to
avoid any strangeness with the tg3 driver. Without this sleep, or with
shorter delays, I have been experiencing TX timeout errors from time to
time. I'm not 100% sure this sleep is a proper fix but I haven't seen TX
timeout errors with it. For now, let's see if it helps.

Finally, this change gets rid of all of our PCI rescan. MicroSemi PCI
port on/off is a much better way to scan the PCI bus.

Change-Id: I08d6d46b8f348ff2bd7538211988c2321f560614