psu: Add PsuSlot and PsuModel abstraction
The PsuSlot class is used to declare a slot as well as detection
mechanism.
A list of PsuModel classes is provided to PsuSlot so that it can
autodetect a compatible power supply from the list.
Once detected the power supply is attached to the slot and initialized
as it was before.
There is caching to avoid performing too much IO and to keep commands
invocation responsive.

Once all the platforms are converted a cleanup will be necessary.

Change-Id: I0463c8ce60f04eea8cff4ffd9e1d8f60173614cd