Tune down arista daemon restart
platform-arista-daemon.service is one of several systemd services that
start as part of SONiC platform initialization. Currently, if platform
initialization fails, then this service gets stuck in a restart loop.
This change modifies the restart behaviour such that the service stops
restarting after a certain number of attempts so that logs don't get
polluted.

Change-Id: I946b36462ca41860e27faf47773bbb1b73564773