provision: Store status using .provision file
When no provision argument is passed to setup --on --lcpu we try to
check wether a /host/provision/ID/.provision file exists. If it does, we
configure provision in static mode. Otherwise, we default to none. Any
command that sets provision to static will create this file. The reverse
is true. If a command explicitly sets provision to none, this file is
removed.

Change-Id: Ie514d9e131ab15cd8d49e481a2e563dbd230f62e