chassis/drv: Don't unconfigure network on driver stop/restart
Tearing down the chassis network bridges during driver stop or restart
is very impactful on the chassis:
- We lose all connections to the linecards (internal & mgmt network)
- The sup doesn't automatically get an IP assisgned to the restored
eth0, leading to the sup also being unreachable

This seems way too disruptive for production. Better do nothing on stop
and make sure we don't do anything on start if everything is already
setup.

In the event where someone wants to remove the internal network, it is
possible to run arista-chassis-network stop manually.

Change-Id: Ia9b5369b8a75de9ae84564641925504638d3f8b8