core/driver: Only rmmod if no device is in use
Currently, when doing a clean of a KernelDriver, we call rmmod to
remove the associated module. This can be problematic if the kernel
module exposes multiple devices. In such a case, the first of those
devices to be cleaned up by the library will result in rmmod of the
module. When going over the rest of the devices for this module, the
library will crash accessing the devices because they are all gone.

This commit addresses this problem by making sure no more devices are
exposed by the module before removing it. Essentially this means that
when a module exposes multiple devices, only the cleanup of the last
one of those devices will result in rmmod being issued.

Change-Id: I45604ad8a168a3087fc0bf1d27e441a2032c1ee0