Add unit test for MetaInventory
A new config option called `use_metainventory` was added.
This review prepares for the transition from Inventory to MetaInventory.
The test currently enables metainventory while still keeping the same
default behavior.
The added value is checking for proper parenting of inventory when used
during platform declaration.
Platforms code was fixed based on this.
Interrupts were refactored to automatically be added to the appropriate
inventory.

Change-Id: I2a4f2f5525d7a703db9fd56d4469839cfcd97b80