Add XcvrSlot abstraction and proper Xcvr types
This change is part of an ongoing effort to refactor xcvr related code
in the platform library.

This adds a slot abstraction for xcvrs similar to the recent effort
to implement a slot abstraction for fans and PSUs, where a slot object
is associated with a corresponding device that is attached to it.
Also adds proper types for each supported xcvr form factor for use in
the platform library, though the platform API will continue to use one
type.

fixes=BUG549963

Change-Id: Idb890d79934609d090f2de8f221a848d7b1a014d