thermal: Limit threshold reads by SfpThermal
SfpThermal depends on Sfp.get_transceiver_threshold_info for obtaining
thermal-related threshold info. Each call does an eeprom read for all of
of the different thresholds, but usage of the thermal API in thermalctld
results in several of these eeprom reads occurring immediately after one
another which is inefficient.

This changes adds a 1 sec delay between reads so that a single read of
the thresholds can be more efficiently used in thermalctld.

Change-Id: I0f3dfbfe3826c6da51ba11b8d5207d97464d07b2