daemon: Add periodic dpm rtc clock sync
The DPM clock might not necessarily be attached to an external crystal.
In such cases and depending on the device the clock will drift.
This will cause reboot cause to have erroneous timestamps.
To prevent this from happening a new state machine will update the clock
of all DPMs in the system every 4h.

Change-Id: I1752ec0df444adadbeaacbcd86cc84a79d177ef3