ports: Cleanup front panel port declaration
This is the final cleanup required for metainventory to work.
Port information used to be stored as variables rather than collections
inside the inventory.

This change also makes more sense as the front panel port layout is
static per platform and it therefore fit nicely as a class declaration.

The next step would be to enhance the PortLayout to have port
characteristics. Something like speeds, lanes, ... which will prove
useful when we get to enhancing ledd.

Change-Id: I5b0d0a73cd918c3ece4516ba0964c9dd2ef6cd12