metainventory: enable single item query
So far the metainventory could only return entire collection that
aggregate data from all the component inventories.
This change also allow query of items by key.
It is relatively safe to do since inventory always use the key to obtain
an item from the inventory and it therefore raises a KeyError when the
item does not exists.

Change-Id: I8e3ce31bcb163aee4eb8bc7db45c2b8878d10f3c