api: refactor get_change_event logic
Move the watching logic outside of chassis.py and make it generic so we
can later reuse it for modules.

The first call on get_change_event will populate all the Events to watch
for and initial statuses.
Upon other iterations of get_change_event the status read will be
compared to the previous status and generate an event if it has changed.

Change-Id: I67f06099b22634ebaf439eefc25be5464b6f8a04