fan-refactor: Add fan support for Lorikeet
The current scd-fan driver assumes 2 fans per fan slot for any board
using a fan platform ID of 3 (woodpecker, lorikeet) and generates
sysfs fans based on this assumption, which leads to phantom fans
being created in cases where there should only be a single fan per
fan slot.

This change introduces a CHASSIS attribute for each platform for holding
chassis-dependent data, like slot count and fan count per slot. This
data is forwarded to the scd-fan driver to generate the correct number
of sysfs fans.
The fan rpm calculation is also enhanced to use an average of the inner
and outer tach readings in the case of a dual rotor fan.

closes: BUG642965

Change-Id: I3506b4bedcabbadf4b91c8eca243c3023874626a