Setup up pylint to show new violations only
The current "make pylint" only runs in --py3k mode and fails due to
violations that have not been fixed and may not be desirable to
fix. Also, the --py3k option is only supported when pylint runs under
Python 2.

Create a new Makefile target "make py3k" to catch Python 3 porting
errors specifically, and make the pylint target generic. Add an
excludes file so that we can exclude checks that currently fail on a
per-file basis. Some of these failures are Python 3 features that we
aren't using due to the need to maintain Python 2 compatibility on
most of the codebase. The remaining violations should be cleaned up at
some point, but by ignoring them for now we ensure that new violations
are caught.

The excludes file is read by a filter script that removes the excluded
checks from the pylint output.

With this change, pylint will not fail for new violations in order to
prevent build issues when this is merged. A future commit will
actually enforce that pylint must pass in order for "make pylint" and
thus "make test" to succeed.

Change-Id: Ic2665ac3ea26d3b65971ce31d489ab9f3ee3e533