tests: Remove testing for use_metainventory=False
The switch to metainventory was done years ago and we no longer need
to support use_metainventory=False. In addition, as it turns out, we
were inadvertently running about half the tests without metainventory
because a test case unconditionally set use_metainventory to False in
its tearDown() method.

With metainventory enabled, we need to use getInventory() whenever we
are accessing the inventory of a component, instead of relying on its
inventory attribute. This fixes a couple of test cases that failed
when the use_metainventory=False was removed.

Change-Id: I0218f502fe142366a413534f9698ac00e60d976a