platform: Add getFanDirectionSku
On PikeZ in the EEPROM the value for SKU
can be either with or without the trailing fan
direction. On newer hardware builds the SKU will be
without the fan direction. In order to determine the
fan direction it has to be parsed from the SID field
in the EEPROM.

In order to maintain consistency that the upper layers see
the SKU with the fan direction, the chassis get_model, get_name
and the syseeprom functions will use getFanDirectionSku
to get the SKU value that has the fan direction appended.

Fixes: BUG881468
Change-Id: I259b27e609aecc7bcfc8490a7bf7ffa80dcfff30
Signed-off-by: Aaron Payment <aaronp@arista.com>